pyproject.toml
[project]
name = "telegram-link-collector"
version = "0.1.0"
description = "Collect links and images from one Telegram chat and optionally sync links to KaraKeep."
readme = "README.md"
authors = [
    { name = "Marc Fabian Mezger", email = "57255687+mfmezger@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "telethon>=1.38.1",
]

[project.scripts]
telegram-link-collector = "telegram_link_collector:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
